# Don't write to tape in numberparsing