This changes number parsing to *return* the number rather than write it immediately to tape. This makes it so we can more freely move tape writing around, and particularly keeps us from having to pass a pointer to the writer to the `never_inline parse_large_integer()` function; if you pass something important like structural_parser, for example, passing it de-optimizes the crap out of it (it is mostly enregistered).

This does not change performance at all on my machine. (In fact, it might actually have given it a roughly 1% bump, but that's possibly noise.)