@jkeiser 2020-05-14
On my machine, g++ 9.3.0, I see no performance difference.
On skylake, g++ 9.2.0, I see a 10% performance degradation.