@jkeiser 2020-05-14
It is in fact the most recent checkin--the one that moves the write index to stage 2--that causes the degradation there. I'm going to change this to remove tape writing from numberparsing, since that moves things forward, and look at that problem separately.