@jkeiser 2020-05-14
skylake gcc-9.2, jkeiser gcc-9.3 and jkeiser clang-6.0 are all within the margin of error.