@jkeiser 2020-05-14
@lemire as best as I can tell, the checkperf failures are noise here. It did find an issue at first, which I addressed. Should be ready to go.

Sadly, I didn't *quite* reach the point of actually moving current_loc from the parser to stage 2 (something I think we want to do with all internals possible, to allow different implementations to vary how they do things). It regressed perf on skylake when I did that, and I haven't investigated yet.