@lemire 2020-05-18
@jkeiser Have you run benchmarks with "number rich" files such as canada.json or numbers.json. Can you publish your results?