@lemire 2020-05-18
Here are my numbers... For master...

```
$ ../buildmaster/benchmark/parse ../jsonexamples/numbers.json
number of iterations 200

../jsonexamples/numbers.json
============================
     2345 blocks -     150124 bytes - 20003 structurals ( 13.3 %)
special blocks with: utf8         0 (  0.0 %) - escape         0 (  0.0 %) - 0 structurals         0 (  0.0 %) - 1+ structurals      2346 (100.0 %) - 8+ structurals      2345 (100.0 %) - 16+ structurals         0 (  0.0 %)
special block flips: utf8         0 (  0.0 %) - escape         0 (  0.0 %) - 0 structurals         0 (  0.0 %) - 1+ structurals         0 (  0.0 %) - 8+ structurals         1 (  0.0 %) - 16+ structurals         0 (  0.0 %)

All Stages
|    Speed        :  66.1266 ns per block ( 98.93%) -   1.0334 ns per byte -   7.7555 ns per structural -   0.9677 GB/s
|    Cycles       : 244.4531 per block    ( 99.03%) -   3.8201 per byte    -  28.6700 per structural    -    3.697 GHz est. frequency
|    Instructions : 737.8291 per block    (100.00%) -  11.5301 per byte    -  86.5344 per structural    -    3.018 per cycle
|    Misses       :    2112 branch misses ( 99.19%) - 0 cache misses (  0.00%) - 19489.00 cache references
|- Stage 1
|    Speed        :  11.0678 ns per block ( 16.56%) -   0.1730 ns per byte -   1.2981 ns per structural -   5.7818 GB/s
|    Cycles       :  40.9702 per block    ( 16.60%) -   0.6402 per byte    -   4.8051 per structural    -    3.702 GHz est. frequency
|    Instructions : 152.4284 per block    ( 20.66%) -   2.3820 per byte    -  17.8772 per structural    -    3.720 per cycle
|    Misses       :      31 branch misses (  1.46%) - 0 cache misses (  0.00%) - 7411.00 cache references
|- Stage 2
|    Speed        :  54.9599 ns per block ( 82.23%) -   0.8589 ns per byte -   6.4458 ns per structural -   1.1643 GB/s
|    Cycles       : 203.0119 per block    ( 82.24%) -   3.1725 per byte    -  23.8097 per structural    -    3.694 GHz est. frequency
|    Instructions : 584.7302 per block    ( 79.25%) -   9.1376 per byte    -  68.5786 per structural    -    2.880 per cycle
|    Misses       :    2075 branch misses ( 97.46%) - 0 cache misses (  0.00%) - 11947.00 cache references

```


For this PR...

```
$ ./benchmark/parse ../jsonexamples/numbers.json 
number of iterations 200 
                                                     
../jsonexamples/numbers.json
============================
     2345 blocks -     150124 bytes - 20003 structurals ( 13.3 %)
special blocks with: utf8         0 (  0.0 %) - escape         0 (  0.0 %) - 0 structurals         0 (  0.0 %) - 1+ structurals      2346 (100.0 %) - 8+ structurals      2345 (100.0 %) - 16+ structurals         0 (  0.0 %)
special block flips: utf8         0 (  0.0 %) - escape         0 (  0.0 %) - 0 structurals         0 (  0.0 %) - 1+ structurals         0 (  0.0 %) - 8+ structurals         1 (  0.0 %) - 16+ structurals         0 (  0.0 %)

All Stages
|    Speed        :  72.5469 ns per block ( 99.68%) -   1.1337 ns per byte -   8.5085 ns per structural -   0.8821 GB/s
|    Cycles       : 268.1159 per block    ( 99.81%) -   4.1899 per byte    -  31.4453 per structural    -    3.696 GHz est. frequency
|    Instructions : 767.7042 per block    (100.00%) -  11.9970 per byte    -  90.0382 per structural    -    2.863 per cycle
|    Misses       :    2012 branch misses ( 99.14%) - 0 cache misses (  0.00%) - 15312.00 cache references
|- Stage 1
|    Speed        :  10.8917 ns per block ( 14.97%) -   0.1702 ns per byte -   1.2774 ns per structural -   5.8752 GB/s
|    Cycles       :  40.3316 per block    ( 15.01%) -   0.6303 per byte    -   4.7302 per structural    -    3.703 GHz est. frequency
|    Instructions : 152.4442 per block    ( 19.86%) -   2.3823 per byte    -  17.8790 per structural    -    3.780 per cycle
|    Misses       :      16 branch misses (  0.79%) - 0 cache misses (  0.00%) - 7057.00 cache references
|- Stage 2
|    Speed        :  61.5512 ns per block ( 84.57%) -   0.9619 ns per byte -   7.2189 ns per structural -   1.0396 GB/s
|    Cycles       : 227.3154 per block    ( 84.62%) -   3.5523 per byte    -  26.6601 per structural    -    3.693 GHz est. frequency
|    Instructions : 614.5891 per block    ( 80.05%) -   9.6042 per byte    -  72.0805 per structural    -    2.704 per cycle
|    Misses       :    1993 branch misses ( 98.21%) - 0 cache misses (  0.00%) - 8065.00 cache references

5875.6 documents parsed per second (best)
```


Let us try with canada.json...

Master:
```
$ ../buildmaster/benchmark/parse ../jsonexamples/canada.json 
number of iterations 200 
                                                     
../jsonexamples/canada.json
===========================
    35172 blocks -    2251051 bytes - 334373 structurals ( 14.9 %)
special blocks with: utf8         0 (  0.0 %) - escape         0 (  0.0 %) - 0 structurals         0 (  0.0 %) - 1+ structurals     35173 (100.0 %) - 8+ structurals     35173 (100.0 %) - 16+ structurals         1 (  0.0 %)
special block flips: utf8         0 (  0.0 %) - escape         0 (  0.0 %) - 0 structurals         0 (  0.0 %) - 1+ structurals         0 (  0.0 %) - 8+ structurals         0 (  0.0 %) - 16+ structurals         2 (  0.0 %)

All Stages
|    Speed        :  71.6046 ns per block ( 99.62%) -   1.1188 ns per byte -   7.5321 ns per structural -   0.8938 GB/s
|    Cycles       : 264.3410 per block    ( 99.72%) -   4.1304 per byte    -  27.8063 per structural    -    3.692 GHz est. frequency
|    Instructions : 916.4568 per block    (100.00%) -  14.3198 per byte    -  96.4029 per structural    -    3.467 per cycle
|    Misses       :   19901 branch misses (100.11%) - 49387 cache misses ( 88.47%) - 324823.00 cache references
|- Stage 1
|    Speed        :  12.8431 ns per block ( 17.87%) -   0.2007 ns per byte -   1.3510 ns per structural -   4.9832 GB/s
|    Cycles       :  47.4223 per block    ( 17.89%) -   0.7410 per byte    -   4.9884 per structural    -    3.692 GHz est. frequency
|    Instructions : 165.6491 per block    ( 18.07%) -   2.5883 per byte    -  17.4248 per structural    -    3.493 per cycle
|    Misses       :    5251 branch misses ( 26.42%) - 10756 cache misses ( 19.27%) - 116138.00 cache references
|- Stage 2
|    Speed        :  58.7188 ns per block ( 81.70%) -   0.9175 ns per byte -   6.1767 ns per structural -   1.0899 GB/s
|    Cycles       : 216.7592 per block    ( 81.77%) -   3.3869 per byte    -  22.8011 per structural    -    3.691 GHz est. frequency
|    Instructions : 750.7630 per block    ( 81.92%) -  11.7308 per byte    -  78.9734 per structural    -    3.464 per cycle
|    Misses       :   14602 branch misses ( 73.46%) - 40567 cache misses ( 72.67%) - 208744.00 cache references

397.1 documents parsed per second (best)
```


This PR:

```
$ ./benchmark/parse ../jsonexamples/canada.json 
number of iterations 200 
                                                     
../jsonexamples/canada.json
===========================
    35172 blocks -    2251051 bytes - 334373 structurals ( 14.9 %)
special blocks with: utf8         0 (  0.0 %) - escape         0 (  0.0 %) - 0 structurals         0 (  0.0 %) - 1+ structurals     35173 (100.0 %) - 8+ structurals     35173 (100.0 %) - 16+ structurals         1 (  0.0 %)
special block flips: utf8         0 (  0.0 %) - escape         0 (  0.0 %) - 0 structurals         0 (  0.0 %) - 1+ structurals         0 (  0.0 %) - 8+ structurals         0 (  0.0 %) - 16+ structurals         2 (  0.0 %)

All Stages
|    Speed        :  76.1494 ns per block ( 99.35%) -   1.1898 ns per byte -   8.0102 ns per structural -   0.8404 GB/s
|    Cycles       : 280.8290 per block    ( 99.35%) -   4.3880 per byte    -  29.5407 per structural    -    3.688 GHz est. frequency
|    Instructions : 943.1794 per block    (100.00%) -  14.7373 per byte    -  99.2139 per structural    -    3.359 per cycle
|    Misses       :   18676 branch misses (100.05%) - 16430 cache misses ( 86.73%) - 278821.00 cache references
|- Stage 1
|    Speed        :  12.7086 ns per block ( 16.58%) -   0.1986 ns per byte -   1.3368 ns per structural -   5.0359 GB/s
|    Cycles       :  46.9210 per block    ( 16.60%) -   0.7331 per byte    -   4.9357 per structural    -    3.692 GHz est. frequency
|    Instructions : 165.6491 per block    ( 17.56%) -   2.5883 per byte    -  17.4248 per structural    -    3.530 per cycle
|    Misses       :    5196 branch misses ( 27.84%) - 6733 cache misses ( 35.54%) - 122909.00 cache references
|- Stage 2
|    Speed        :  63.3906 ns per block ( 82.70%) -   0.9905 ns per byte -   6.6681 ns per structural -   1.0096 GB/s
|    Cycles       : 233.7182 per block    ( 82.68%) -   3.6519 per byte    -  24.5850 per structural    -    3.687 GHz est. frequency
|    Instructions : 777.4855 per block    ( 82.43%) -  12.1483 per byte    -  81.7844 per structural    -    3.327 per cycle
|    Misses       :   13405 branch misses ( 71.81%) - 9198 cache misses ( 48.55%) - 155865.00 cache references
```

So there is a pretty clear cut regression on number-rich files. 


The numbers I am getting would be consistent with a model where the number of instructions per number parsed is up by 7 to 9 instructions. That is, the new code would require 7 to 9 more instructions per number parsed.


And these instructions nearly cost you up to five cycles per number parsed.